third_party/blink/web_tests/external/wpt/content-security-policy/webrtc/webrtc-blocked-explicit-expected.txt

This is a testharness.js-based test.
[FAIL] webrtc blocked with an explicit webrtc blocked policy
  assert_equals: expected "blocked" but got "allowed"
Harness: the test ran to completion.
